requirements.txt
altair==5.5.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.55.3
gitdb==4.0.12
GitPython==3.1.44
graphviz==0.20.3
greenlet==3.1.1
idna==3.10
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
narwhals==1.30.0
numpy==2.2.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pi==0.1.2
pillow==11.1.0
plotly==6.0.1
plotly-express==0.4.1
protobuf==5.29.3
# psycopg2==2.9.10
psycopg2-binary==2.9.10
pyarrow==19.0.1
pydeck==0.9.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
referencing==0.36.2
reportlab==4.3.1
requests==2.32.3
rpds-py==0.23.1
scipy==1.15.2
seaborn==0.13.2
setuptools==76.0.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.39
statsmodels==0.14.4
streamlit==1.43.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tz==0.2.2
tzdata==2024.2
urllib3==2.3.0
watchdog==6.0.0
wheel==0.45.1
XlsxWriter==3.2.2
